Promote standard token-exchange V2 to supported by default closes #37368
